지금까지 한 작업
